requirements.txt
flask==3.0.0
flask-cors==4.0.0
google-generativeai==0.5.0
uuid==1.30
python-dotenv==1.0.1
requests==2.31.0 